travel-planner-agent-demo/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pyautogen>=0.2.0
yahooquery>=2.3.10